// src/health-answers/health-answers.enums.ts
export enum HealthAnswerType {
  ENGINE_OIL = 'engine_oil',
  BRAKES = 'brakes',
  TIRES = 'tires',
  TIMING_BELT = 'timing_belt',
  COOLING_SYSTEM = 'cooling_system',
}

export enum HealthAnswerValue {
  RECENT = 'recent',
  WHILE = 'while',
  OVER_YEAR = 'over_year',
  NEVER = 'never',
  NO_TIMING_BELT = 'no_timing_belt',
}